requirements.txt
opencv-python
torchvision
torch
matplotlib
numpy
pillow
requests